Fixing really slow page loads
Whenever a category or tag page gets loaded that includes the comment count, a zillion requests end up going off to sync the comments for every post on the page that hasn't been viewed recently, making things really slow. This change limits the sync comments to only when the post itself is viewed.